requirements.txt
asgiref==3.5.0
Django==4.0.2
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
